+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/data/dto/StatLegendDto.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class StatLegendDto(
     @SerialName("legend_id")
-    val legendId: Int,
+    val legendId: Long,
     @SerialName("legend_name_key")
     val legendNameKey: String,
     @SerialName("damagedealt")
