+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/domain/RankingLegend.kt
@@ -1,7 +1,7 @@
 package com.nickoehler.brawlhalla.ranking.domain
 
 data class RankingLegend(
-    val legendId: Int,
+    val legendId: Long,
     val legendNameKey: String,
     val rating: Int,
     val peakRating: Int,
